refactor(formbuilder): rename FormBuilder and added ValidForm Builder 3 compatibility
