+++ classes/PunchCMS/Client/Client.php
@@ -914,7 +914,7 @@ class Client
 			    PDO::ATTR_PERSISTENT => true
 			));
 		} catch (PDOException $e) {
-			throw new Exception('Database connection failed: ' . $e->getMessage(), SQL_CONN_ERROR);
+			throw new \Exception('Database connection failed: ' . $e->getMessage(), SQL_CONN_ERROR);
 		}
 
 		self::$__connId = $objConnID;
